+++ OUDS/Core/Components/Sources/_/ViewModifiers/TypographyModifiers/TypographyModifier.swift
@@ -22,7 +22,7 @@ import UIKit
 // swiftlint:disable line_length
 
 /// A `ViewModifier` which will make possible to get the horizontal and vertical classes as `@Environment` values
-/// so as to define the viewport and use finaly the suitable `MultipleFontCompositeRawTokens`.
+/// so as to define the viewport and use finaly the suitable `MultipleFontCompositeSemanticTokens`.
 /// In fact dimensionswift extension_ does not allow to have such stored properties, and we don't want to use *UIKit* `UIScreen.main.traitCollection` to get values which may be out of date.
 /// In few words, contains the font elements to apply a defined typography depending to size classes and categories.
 /// For more details about layouts, see [the Apple documentation about devices dimensions](https://developer.apple.com/design/human-interface-guidelines/layout#iOS-iPadOS-device-size-classes)
@@ -32,7 +32,7 @@ struct TypographyModifier: ViewModifier {
     let family: FontFamilyRawToken?
 
     /// The typography to apply for *compact* or *regular* modes, i.e. font tokens
-    let font: MultipleFontCompositeRawTokens
+    let font: MultipleFontCompositeSemanticTokens
 
     /// To get programatically and on the fly the horizontal layout size
     @Environment(\.horizontalSizeClass) private var horizontalSizeClass
@@ -50,7 +50,7 @@ struct TypographyModifier: ViewModifier {
     }
 
     /// Returns the `FontCompositeRawToken` to apply depending to the layour mode
-    private var adaptiveFont: FontCompositeRawToken {
+    private var adaptiveFont: FontCompositeSemanticToken {
         isCompactMode ? font.compact : font.regular
     }
 
@@ -82,7 +82,7 @@ struct TypographyModifier: ViewModifier {
     }
 
     /// Applies to the `Content` the *adaptive font* (i.e. *font family*, *font weight* and *font size*
-    /// depending to the current `MultipleFontCompositeRawTokens`.
+    /// depending to the current `MultipleFontCompositeSemanticTokens`.
     func body(content: Content) -> some View {
         // `tracking()` only available for iOS 16+
         // `minimumScaleFactor()` ensures text remains readable by allowing scaling down
@@ -138,7 +138,7 @@ struct TypographyModifier: ViewModifier {
     }
 
     /// Applies to the `Content` the *adaptive font* (i.e. *font family*, *font weight* and *font size*
-    /// depending to the current `MultipleFontCompositeRawTokens`.
+    /// depending to the current `MultipleFontCompositeSemanticTokens`.
     func body(content: Content) -> some View {
         // `tracking()` only available for iOS 16+
         // `minimumScaleFactor()` ensures text remains readable by allowing scaling down
@@ -152,7 +152,7 @@ struct TypographyModifier: ViewModifier {
     }
     #else
     /// Applies to the `Content` the *adaptive font* (i.e. *font family*, *font weight* and *font size*
-    /// depending to the current `MultipleFontCompositeRawTokens`.
+    /// depending to the current `MultipleFontCompositeSemanticTokens`.
     func body(content: Content) -> some View {
         // `tracking()` only available for iOS 16+
         // `minimumScaleFactor()` ensures text remains readable by allowing scaling down