+++ OUDS/Core/Tokens/SemanticTokens/Sources/Values/FontCompositeSemanticToken.swift
@@ -11,7 +11,9 @@
 // Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
-/// An operator to make for example comparisons between ``FontCompositeRawToken``
+import OUDSTokensRaw
+
+/// An operator to make for example comparisons between ``FontCompositeSemanticToken``
 infix operator <|
 
 /// In the *Figma* global design system, composite tokens are defined for font-related things.
@@ -19,8 +21,8 @@ infix operator <|
 /// Here a *font* thing is caracterized by a *font size*, a *line height*, a *font weight* and *letter spacing* values.
 /// All these elements are *raw tokens*, and together define a *composite raw token* for *font* thing.
 ///
-/// - Since: 0.8.0
-public struct FontCompositeRawToken: Equatable, Sendable {
+/// - Since: 0.22.0
+public struct FontCompositeSemanticToken: Equatable, Sendable {
 
     // Font family is not included here because this is the only thing which can vary
 
@@ -36,6 +38,22 @@ public struct FontCompositeRawToken: Equatable, Sendable {
     /// The font letter spacing to associated with the font family
     public let letterSpacing: FontLetterSpacingRawToken
 
+    /// - Parameters:
+    ///    - size: the size for the texts
+    ///    - lineHeight: the line height for the texts
+    ///    - weight: the weight for the texts
+    ///    - letterSpacing: the letter spacing for the texts
+    public init(size: FontSizeRawToken,
+                lineHeight: FontLineHeightRawToken,
+                weight: FontWeightRawToken,
+                letterSpacing: FontLetterSpacingRawToken)
+    {
+        self.size = size
+        self.lineHeight = lineHeight
+        self.weight = weight
+        self.letterSpacing = letterSpacing
+    }
+
     /// Operator which will return `true` if `lhs` is smaller than `rhs`.
     /// By "smaller" we mean smaller `size` and smaller or equal `lineHeight` and `weight`.
     ///
@@ -47,7 +65,7 @@ public struct FontCompositeRawToken: Equatable, Sendable {
     ///    - lhs: The font composite token we expect to be smaller than `rhs`
     ///    - rhs: The font composite token we expect to be bigger than `lhs`
     /// - Returns Bool: `true` if `lhs` smaller than `rhs`, `false` otherwise
-    public static func <| (lhs: FontCompositeRawToken, rhs: FontCompositeRawToken) -> Bool {
+    public static func <| (lhs: FontCompositeSemanticToken, rhs: FontCompositeSemanticToken) -> Bool {
         lhs.size < rhs.size
             && lhs.lineHeight <= rhs.lineHeight
             && lhs.weight <= rhs.weight